# src/backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CapitalForge API
  version: 2.0.0
  description: |
    CapitalForge is a multi-tenant B2B lending-stack platform that guides small
    businesses from credit profiling through card-stack optimisation, funding
    rounds, compliance, billing, and AI-assisted document intelligence.

    **Authentication** — All protected endpoints require a Bearer JWT obtained
    from `POST /api/auth/login`. Include it as:
    ```
    Authorization: Bearer <accessToken>
    ```
    Tokens expire after 15 minutes. Use `POST /api/auth/refresh` to rotate.

    **Tenant scoping** — Every resource is isolated by `tenantId` encoded in
    the JWT. Passing an `X-Tenant-Id` header is also accepted in dev/test mode.

  contact:
    name: CapitalForge Platform Engineering
    email: platform@capitalforge.io
  license:
    name: Proprietary
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://staging-api.capitalforge.io/api
    description: Staging
  - url: https://api.capitalforge.io/api
    description: Production
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# Tags
# ─────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT. Obtain via POST /api/auth/login.
  parameters:
    businessId:
      name: id
      in: path
      required: true
      description: UUID of the Business resource
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        data:
          description: Present on success; shape varies by endpoint
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request body.
        details:
          description: Field-level validation errors or additional context
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer

    # ── Auth ────────────────────────────────────────────────────────────────
...
    CreditProfile:
      type: object
      required: [id, businessId, bureau, score, pulledAt]
      properties:
        id:
          type: string
          format: uuid
        businessId:
          type: string
          format: uuid
        bureau:
          type: string
          enum: [experian, equifax, transunion, dun_bradstreet, lexisnexis]
        profileType:
          type: string
          enum: [personal, business]
        score:
          type: integer
        factors:
          type: array
          items:
            type: string
        utilizationRate:
          type: number
        derogatoryMarks:
          type: integer
        synthetic:
          type: boolean
          description: >-
            Whether these figures came from a bureau or were generated. Derived
            from rawData.synthetic on read; there is no column behind it.
        pulledAt:
          type: string
          format: date-time
    CreditPullRequest:
      type: object
      required: [bureau, profileType]
      properties:
        # One bureau per call. This was `bureaus: Bureau[]` until 7 September
        # 2026, when the array made a partial success representable: 201 with
        # fewer profiles than bureaus requested, and a total counting successes.
        bureau:
          type: string
          enum: [experian, equifax, transunion, dun_bradstreet, lexisnexis]
        profileType:
          type: string
          enum: [personal, business]
        useCache:
          type: boolean
          default: false

    # ── Card Application ─────────────────────────────────────────────────────
...
paths:
...
  /businesses/{id}/credit:
    get:
      tags: [Core Operations]
      operationId: getCreditProfiles
      summary: List all credit profiles for a business
      parameters:
        - $ref: '#/components/parameters/businessId'
      responses:
        '200':
          description: Credit profiles list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/CreditProfile'
  /businesses/{id}/credit/pull:
    post:
      tags: [Core Operations]
      operationId: pullCreditProfiles
      summary: Trigger a fresh bureau credit pull
      parameters:
        - $ref: '#/components/parameters/businessId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditPullRequest'
      responses:
        '201':
          description: The profile the pull produced. One bureau, one profile.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CreditProfile'
        '503':
          description: >-
            The bureau is not configured. Not a 500: the request was fine and the
            integration is unavailable, and a caller that retries a 500 is
            retrying an attempt to write an invented credit profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /businesses/{id}/credit/roadmap:
    get:
      tags: [Core Operations]
      operationId: getCreditRoadmap
      summary: Get a prioritised credit optimisation roadmap
      parameters:
        - $ref: '#/components/parameters/businessId'
      responses:
        '200':
          description: Optimisation roadmap
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ── Funding Rounds ─────────────────────────────────────────────────────────
...